requirements.txt
# LiveKit Agent SDK
livekit>=0.18.0
livekit-agents>=0.12.0
livekit-plugins-google>=0.10.0
livekit-plugins-silero>=0.7.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Google Gemini AI (for Computer Use)
google-genai>=1.0.0

# Browser Automation (for Gemini Computer Use)
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# For desktop automation (optional)
pyautogui>=0.9.54
